Move wallet unlock dialog store
